[Bug 1816150]: Can't replace dir with file.
If a managed object should be a file but is a directory on the disk,
causing a "+ f ./path/to/file ..." transcript line to be written out,
AND the next line in the appliable transcript is a transcript context
change ("transcript-name.T:"), previous versions of lapply would use
the new, incorrect transcript context when requesting the missing file
from radmind. The fix adds tracking of the transcript context to the
node struct.